+++ crates/punktf-lib/src/action/deploy/visit.rs
@@ -1,204 +1,88 @@
 use color_eyre::eyre::Context;
 
-use crate::template::Template;
-use crate::{profile::visit::*, template::source::Source};
+use crate::profile::visit::*;
 use crate::{MergeMode, PunktfSource};
 
+use crate::action::deploy::{Deployment, DeploymentBuilder, DotfileStatus};
 use crate::profile::LayeredProfile;
 use crate::transform::Transform as _;
 use std::borrow::Borrow;
 use std::path::Path;
 
-use super::deployment::{Deployment, DeploymentBuilder};
-use super::dotfile::DotfileStatus;
-use super::executor::ExecutorOptions;
+use crate::profile::visit::{ResolvingVisitor, TemplateVisitor};
 
-pub trait TemplateVisitor: Visitor {
-	fn accept_template<'a>(
-		&mut self,
-		source: &PunktfSource,
-		profile: &LayeredProfile,
-		file: &File<'a>,
-		// Returns a function to resolve the content to make the resolving lazy
-		// for upstream visitors.
-		resolve_content: impl FnOnce(&str) -> color_eyre::Result<String>,
-	) -> Result;
-}
+macro_rules! add_to_builder {
+	($item:ty) => {
+		impl $item {
+			fn add_to_builder<S: Into<DotfileStatus>>(
+				&self,
+				builder: &mut DeploymentBuilder,
+				status: S,
+			) {
+				let status = status.into();
 
-#[derive(Debug)]
-pub struct ResolvingVisitor<V> {
-	visitor: V,
-}
+				let resolved_target_path = self
+					.target_path
+					.canonicalize()
+					.unwrap_or_else(|_| self.target_path.clone());
 
-impl<V> ResolvingVisitor<V>
-where
-	V: TemplateVisitor,
-{
-	pub fn new(visitor: V) -> Self {
-		Self { visitor }
-	}
+				match &self.kind {
+					Kind::Root(dotfile) => {
+						builder.add_dotfile(resolved_target_path, (*dotfile).clone(), status)
+					}
+					Kind::Child {
+						root_target_path, ..
+					} => {
+						let resolved_root_target_path = root_target_path
+							.canonicalize()
+							.unwrap_or_else(|_| root_target_path.clone());
 
-	pub fn into_inner(self) -> V {
-		self.visitor
-	}
-}
-
-impl<V: TemplateVisitor> Visitor for ResolvingVisitor<V> {
-	fn accept_file<'a>(
-		&mut self,
-		source: &PunktfSource,
-		profile: &LayeredProfile,
-		file: &File<'a>,
-	) -> Result {
-		println!("F: {:#?}", file);
-
-		if file.dotfile().is_template() {
-			let resolve_fn = |content: &str| {
-				let source = Source::file(&file.source_path, content);
-				let template = Template::parse(source)
-					.with_context(|| format!("File: {}", file.source_path.display()))?;
-
-				template
-					.resolve(Some(profile.variables()), file.dotfile().variables.as_ref())
-					.with_context(|| format!("File: {}", file.source_path.display()))
-			};
-
-			self.visitor
-				.accept_template(source, profile, file, resolve_fn)
-		} else {
-			self.visitor.accept_file(source, profile, file)
+						builder.add_child(resolved_target_path, resolved_root_target_path, status)
+					}
+				};
+			}
 		}
-	}
-
-	fn accept_directory<'a>(
-		&mut self,
-		source: &PunktfSource,
-		profile: &LayeredProfile,
-		directory: &Directory<'a>,
-	) -> Result {
-		println!("D: {:#?}", directory);
-
-		self.visitor.accept_directory(source, profile, directory)
-	}
-
-	fn accept_link(
-		&mut self,
-		source: &PunktfSource,
-		profile: &LayeredProfile,
-		symlink: &Symlink,
-	) -> Result {
-		self.visitor.accept_link(source, profile, symlink)
-	}
-
-	fn accept_rejected<'a>(
-		&mut self,
-		source: &PunktfSource,
-		profile: &LayeredProfile,
-		rejected: &Rejected<'a>,
-	) -> Result {
-		self.visitor.accept_rejected(source, profile, rejected)
-	}
-
-	fn accept_errored<'a>(
-		&mut self,
-		source: &PunktfSource,
-		profile: &LayeredProfile,
-		errored: &Errored<'a>,
-	) -> Result {
-		self.visitor.accept_errored(source, profile, errored)
-	}
+	};
 }
 
-impl<'a> File<'a> {
-	fn add_to_builder<S: Into<DotfileStatus>>(&self, builder: &mut DeploymentBuilder, status: S) {
-		let status = status.into();
+add_to_builder!(File<'_>);
+add_to_builder!(Directory<'_>);
+add_to_builder!(Rejected<'_>);
 
-		let resolved_target_path = self
-			.target_path
-			.canonicalize()
-			.unwrap_or_else(|_| self.target_path.clone());
-
-		match &self.kind {
-			Kind::Root(dotfile) => {
-				builder.add_dotfile(resolved_target_path, (*dotfile).clone(), status)
-			}
-			Kind::Child {
-				root_target_path, ..
-			} => {
-				let resolved_root_target_path = root_target_path
-					.canonicalize()
-					.unwrap_or_else(|_| root_target_path.clone());
-
-				builder.add_child(resolved_target_path, resolved_root_target_path, status)
-			}
-		};
-	}
+/// Configuration options for the [`DeployingVisitor`].
+#[derive(Default, Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub struct DeployOptions {
+	/// If this flag is set, it will prevent any write operations from occurring
+	/// during the deployment.
+	///
+	/// This includes write, copy and directory creation operations.
+	pub dry_run: bool,
 }
 
-impl<'a> Directory<'a> {
-	fn add_to_builder<S: Into<DotfileStatus>>(&self, builder: &mut DeploymentBuilder, status: S) {
-		let status = status.into();
-
-		let resolved_target_path = self
-			.target_path
-			.canonicalize()
-			.unwrap_or_else(|_| self.target_path.clone());
-
-		match &self.kind {
-			Kind::Root(dotfile) => {
-				builder.add_dotfile(resolved_target_path, (*dotfile).clone(), status)
-			}
-			Kind::Child {
-				root_target_path, ..
-			} => {
-				let resolved_root_target_path = root_target_path
-					.canonicalize()
-					.unwrap_or_else(|_| root_target_path.clone());
-
-				builder.add_child(resolved_target_path, resolved_root_target_path, status)
-			}
-		};
-	}
-}
-
-impl<'a> Rejected<'a> {
-	fn add_to_builder<S: Into<DotfileStatus>>(&self, builder: &mut DeploymentBuilder, status: S) {
-		let status = status.into();
-
-		let resolved_target_path = self
-			.target_path
-			.canonicalize()
-			.unwrap_or_else(|_| self.target_path.clone());
-
-		match &self.kind {
-			Kind::Root(dotfile) => {
-				builder.add_dotfile(resolved_target_path, (*dotfile).clone(), status)
-			}
-			Kind::Child {
-				root_target_path, ..
-			} => {
-				let resolved_root_target_path = root_target_path
-					.canonicalize()
-					.unwrap_or_else(|_| root_target_path.clone());
-
-				builder.add_child(resolved_target_path, resolved_root_target_path, status)
-			}
-		};
-	}
-}
-
+/// Responsible for deploying a [profile](`crate::profile::Profile`).
+///
+/// This includes checking for merge conflicts, resolving children of a
+/// directory dotfile, parsing and resolving of templates and the actual
+/// writing of the dotfile to the target destination.
 #[derive(Debug, Clone, PartialEq, Eq)]
-pub struct DeployingVisitor<F> {
-	options: ExecutorOptions,
+pub struct Deployer<F> {
+	/// Configuration options
+	options: DeployOptions,
+
+	/// This function gets called when a dotfile at the target destination
+	/// already exists and the merge mode is
+	/// [MergeMode::Ask](`crate::MergeMode::Ask`).
+	///
+	/// The arguments for the function are (dotfile_source_path, dotfile_target_path).
 	merge_ask_fn: F,
 	builder: DeploymentBuilder,
 }
 
-impl<F> DeployingVisitor<F>
+impl<F> Deployer<F>
 where
 	F: Fn(&Path, &Path) -> color_eyre::Result<bool>,
 {
-	pub fn new(options: ExecutorOptions, merge_ask_fn: F) -> Self {
+	pub fn new(options: DeployOptions, merge_ask_fn: F) -> Self {
 		Self {
 			options,
 			merge_ask_fn,
@@ -208,6 +92,68 @@ where
 
 	pub fn into_deployment(self) -> Deployment {
 		self.builder.finish()
+	}
+
+	/// Tries to deploy the given `profile`.
+	///
+	/// # Errors
+	///
+	/// Only hard errors will be returned as error, everthing else will be
+	/// recorded in the [Deployment](`super::deployment::Deployment`) on a
+	/// dotfile level.
+	pub fn deploy(self, source: &PunktfSource, profile: LayeredProfile) -> Deployment {
+		// General flow:
+		//	- get deployment path
+		//	- check if dotfile already deployed
+		//	- YES:
+		//		- compare priorities
+		//		- LOWER: continue next dotfile
+		//		- SAME/HIGHER: next step
+		//	- check if dotfile exists
+		//	- YES:
+		//		- check merge operation
+		//		- if merge operation == ASK
+		//			- Run merge_ask_fn
+		//			- FALSE: continue next dotfile
+		//	- check if template
+		//	- YES: resolve template
+		//	- IF FILE: write dotfile
+		//	- IF DIR: for each dotfile in dir START AT TOP
+
+		for hook in profile.pre_hooks() {
+			log::info!("Executing pre-hook: {}", hook.command());
+			// No files are deployed yet, meaning if an error during hook
+			// execution occurs it will return with an error instead of just
+			// logging it.
+
+			if let Err(err) = hook
+				.execute(source.profiles())
+				.wrap_err("Failed to execute pre-hook")
+			{
+				log::error!("Failed to execute pre-hook ({})", err);
+				return self.builder.failed(err.to_string());
+			};
+		}
+
+		let post_hooks: Vec<_> = profile.post_hooks().cloned().collect();
+
+		let mut resolver = ResolvingVisitor::new(self);
+		let walker = Walker::new(profile);
+		if let Err(err) = walker.walk(source, &mut resolver) {
+			return resolver.into_inner().builder.failed(err.to_string());
+		}
+
+		let this = resolver.into_inner();
+
+		for hook in post_hooks {
+			log::info!("Executing post-hook: {}", hook.command());
+			if let Err(err) = hook.execute(source.profiles()) {
+				log::error!("Failed to execute post-hook ({})", err);
+				return this.builder.failed(err.to_string());
+			}
+		}
+
+		this.into_deployment()
 	}
 
 	fn pre_deploy_checks(
@@ -225,9 +171,9 @@ where
 				)
 			})?;
 
-		let other_priority = self.builder.get_priority(&file.target_path).unwrap_or(None);
+		let other_priority = self.builder.get_priority(&file.target_path);
 
-		match (file.dotfile().priority.as_ref(), other_priority.as_ref()) {
+		match (file.dotfile().priority.as_ref(), other_priority) {
 			(Some(a), Some(b)) if b > a => {
 				log::info!(
 					"{}: Dotfile with higher priority is already deployed at {}",
@@ -392,7 +338,7 @@ where
 	}
 }
 
-impl<F> Visitor for DeployingVisitor<F>
+impl<F> Visitor for Deployer<F>
 where
 	F: Fn(&Path, &Path) -> color_eyre::Result<bool>,
 {
@@ -556,7 +502,7 @@ where
 	}
 }
 
-impl<F> TemplateVisitor for DeployingVisitor<F>
+impl<F> TemplateVisitor for Deployer<F>
 where
 	F: Fn(&Path, &Path) -> color_eyre::Result<bool>,
 {
