--- crates/punktf-lib/src/deploy/mod.rs
@@ -1,6 +1,0 @@
-//! Everting needed to deploy a [profile](`crate::profile::Profile`).
-
-pub mod deployment;
-pub mod dotfile;
-pub mod executor;
-pub mod visit;
